@@ -13,7 +13,7 @@
     "@stomp/stompjs": "^5.3.0",
     "bootstrap": "^4.3.1",
     "history": "^4.9.0",
-    "react": "^16.8.4",
+    "react": "16.13.1",
     "react-bootstrap": "^1.0.0-beta.5",
     "react-dom": "^16.8.4",
     "react-redux": "^6.0.1",
