@@ -8,43 +8,33 @@ interface IFindByNameProps {
     getUserByName: (userName: string) => any;
 }
 
-interface IFindByNameState {
-    users: IUser[];
-}
+export const FindByName = (props: IFindByNameProps) => {
 
-export class FindByName extends React.Component<IFindByNameProps, IFindByNameState> {
+    const [users, setUsers] = React.useState<IUser[]>([]);
 
-    public state: any = {
-        users: [],
+    const findUser = (userName: string) => {
+        if (userName) {
+            props.getUserByName(userName).then((users: IUser[]) => setUsers(users));
+        }
     };
 
-    private findUser = (userName: string) => {
-        if (userName) {
-            this.props.getUserByName(userName).then((users: IUser[]) => this.setState({ users }));
-        }
-    }
-
-    public render() {
-        const options: IUser[] = this.state.users;
-
-        return (
-            <div>
-                <Autocomplete
-                    freeSolo
-                    style={ { width: 300 } }
-                    options={ options }
-                    onInputChange={ (event, newInputValue) => {
-                        if (event.type === 'change') {
-                            this.findUser(newInputValue);
-                        }
-                    } }
-                    renderOption={ (option: IUser) => (
-                        <Link to={ `/profile/${ option.id }` }>{ option.name }</Link>
-                    ) }
-                    getOptionLabel={ (option: IUser) => option.name }
-                    renderInput={ (params) =>  <TextField { ...params } label="Find user" variant="outlined" /> }
-                />
-            </div>
-        );
-    }
-}
+    return (
+        <div>
+            <Autocomplete
+                freeSolo
+                style={ { width: 300 } }
+                options={ users }
+                onInputChange={ (event, newInputValue) => {
+                    if (event.type === 'change') {
+                        findUser(newInputValue);
+                    }
+                } }
+                renderOption={ (option: IUser) => (
+                    <Link to={ `/profile/${ option.id }` }>{ option.name }</Link>
+                ) }
+                getOptionLabel={ (option: IUser) => option.name }
+                renderInput={ (params) => <TextField { ...params } label="Find user" variant="outlined"/> }
+            />
+        </div>
+    );
+};
